+++ src/screens/VisualTests/VisualTests.stories.tsx
@@ -128,12 +128,13 @@ const withGraphQLQuery = (...args: Parameters<typeof graphql.query>) => ({
   },
 });
 
-function withGraphQLQueryResult<TQuery extends TypedDocumentNode>(
+function withGraphQLQueryResult<TQuery extends TypedDocumentNode<any, any>>(
   query: TQuery,
   result: ResultOf<TQuery>
 ) {
-  const queryName = getOperationAST(query).name.value;
-  return withGraphQLQuery(queryName, (req, res, ctx) => res(ctx.data(result)));
+  const queryName = getOperationAST(query)?.name?.value;
+  if (queryName) return withGraphQLQuery(queryName, (req, res, ctx) => res(ctx.data(result)));
+  throw new Error(`Couldn't determine query name from query`);
 }
 
 const withGraphQLMutation = (...args: Parameters<typeof graphql.mutation>) => ({
@@ -195,7 +196,7 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Loading: Story = {
+export const Loading = {
   parameters: {
     ...withGraphQLQuery("AddonVisualTestsBuild", (req, res, ctx) =>
       res(ctx.status(200), ctx.data({}), ctx.delay("infinite"))
@@ -204,19 +205,19 @@ export const Loading: Story = {
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=508-304933&t=0rxMQnkxsVpVj1qy-4"
     ),
   },
-};
+} satisfies Story;
 
-export const GraphQLError: Story = {
+export const GraphQLError = {
   parameters: {
     ...withGraphQLQuery("AddonVisualTestsBuild", (req, res, ctx) =>
       res(ctx.status(200), ctx.errors([{ message: "Something went wrong on the server" }]))
     ),
   },
-};
+} satisfies Story;
 
-export const EmptyBranch: Story = {
+export const EmptyBranch = {
   parameters: {
-    ...withBuilds({ selectedBuild: null }),
+    ...withBuilds({ selectedBuild: undefined }),
   },
   render: ({ ...args }) => {
     // custom render for mapping `updateBuildStatus` to a function which is mocked, but returns data instead of a function
@@ -232,9 +233,9 @@ export const EmptyBranch: Story = {
       expect(args.updateBuildStatus).toHaveBeenCalledWith({});
     });
   },
-};
+} satisfies Story;
 
-export const EmptyBranchStartedLocalBuild: Story = {
+export const EmptyBranchStartedLocalBuild = {
   ...EmptyBranch,
   args: {
     localBuildProgress: {
@@ -246,9 +247,9 @@ export const EmptyBranchStartedLocalBuild: Story = {
       },
     },
   },
-};
+} satisfies Story;
 
-export const EmptyBranchLocalBuildUploading: Story = {
+export const EmptyBranchLocalBuildUploading = {
   ...EmptyBranch,
   args: {
     localBuildProgress: {
@@ -265,12 +266,12 @@ export const EmptyBranchLocalBuildUploading: Story = {
       },
     },
   },
-};
+} satisfies Story;
 
 /** This story should maintain the "no build" UI with a progress bar */
-export const EmptyBranchLocalBuildCapturing: Story = {
+export const EmptyBranchLocalBuildCapturing = {
   parameters: {
-    ...withBuilds({ selectedBuild: null, lastBuildOnBranch: inProgressBuild }),
+    ...withBuilds({ selectedBuild: undefined, lastBuildOnBranch: inProgressBuild }),
   },
   args: {
     localBuildProgress: {
@@ -287,13 +288,13 @@ export const EmptyBranchLocalBuildCapturing: Story = {
       },
     },
   },
-};
+} satisfies Story;
 
 /** At this point, we should switch to the next build */
-export const EmptyBranchLocalBuildCapturedCurrentStory: Story = {
+export const EmptyBranchLocalBuildCapturedCurrentStory = {
   parameters: {
     ...withBuilds({
-      selectedBuild: null,
+      selectedBuild: undefined,
       lastBuildOnBranch: withTests(
         inProgressBuild,
         SnapshotComparisonStories.WithSingleTest.args.tests
@@ -313,27 +314,27 @@ export const EmptyBranchLocalBuildCapturedCurrentStory: Story = {
       },
     },
   },
-};
+} satisfies Story;
 
 /** Complete builds should always be switched to */
-export const EmptyBranchCIBuildPending: Story = {
+export const EmptyBranchCIBuildPending = {
   parameters: {
     ...withBuilds({
-      selectedBuild: null,
+      selectedBuild: undefined,
       lastBuildOnBranch: pendingBuild,
     }),
   },
   // In theory we might have a complete running build here, it should behave the same either way
-};
+} satisfies Story;
 
-export const NoChanges: Story = {
+export const NoChanges = {
   parameters: {
     ...withBuilds({ selectedBuild: passedBuild }),
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=508-304933&t=0rxMQnkxsVpVj1qy-4"
     ),
   },
-};
+} satisfies Story;
 
 /** We just switched branches so the selected build is out of date */
 export const NoChangesOnWrongBranch: Story = {
@@ -351,19 +352,19 @@ export const NoChangesOnWrongBranch: Story = {
 /**
  * We've started a new build but it's not done yet
  */
-export const PendingLocalBuildStarting: Story = {
+export const PendingLocalBuildStarting = {
   args: {
     ...EmptyBranchStartedLocalBuild.args,
   },
   parameters: {
     ...withBuilds({ selectedBuild: pendingBuild }),
   },
-};
+} satisfies Story;
 
 /**
  * As above but we started the next build
  */
-export const PendingLocalBuildCapturing: Story = {
+export const PendingLocalBuildCapturing = {
   parameters: {
     ...withBuilds({
       selectedBuild: pendingBuild,
@@ -373,12 +374,12 @@ export const PendingLocalBuildCapturing: Story = {
   args: {
     ...EmptyBranchLocalBuildCapturing.args,
   },
-};
+} satisfies Story;
 
 /**
  * The next build is snapshotting and has captured this story
  */
-export const PendingLocalBuildCapturedStory: Story = {
+export const PendingLocalBuildCapturedStory = {
   ...PendingLocalBuildCapturing,
   parameters: {
     ...withBuilds({
@@ -392,28 +393,28 @@ export const PendingLocalBuildCapturedStory: Story = {
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=2303-374529&t=qjmuGHxoALrVuhvX-0"
     ),
   },
-};
+} satisfies Story;
 
 /**
  * The next build is snapshotting but hasn't yet reached this story (we didn't start it)
  */
-export const PendingCIBuildInProgress: Story = {
+export const PendingCIBuildInProgress = {
   parameters: PendingLocalBuildCapturing.parameters,
-};
+} satisfies Story;
 
 /**
  * The next build is snapshotting and has captured this story
  */
-export const PendingCIBuildCapturedStory: Story = {
+export const PendingCIBuildCapturedStory = {
   parameters: {
     ...PendingLocalBuildCapturedStory.parameters,
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=2303-374529&t=qjmuGHxoALrVuhvX-0"
     ),
   },
-};
+} satisfies Story;
 
-export const Pending: Story = {
+export const Pending = {
   parameters: {
     ...withBuilds({ selectedBuild: pendingBuild }),
     ...withFigmaDesign(
@@ -440,36 +441,36 @@ export const Pending: Story = {
       });
     });
   },
-};
+} satisfies Story;
 
-export const NoPermission: Story = {
+export const NoPermission = {
   parameters: {
     ...withBuilds({ selectedBuild: pendingBuild, userCanReview: false }),
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=2127-449276&mode=design&t=gIM40WT0324ynPQD-4"
     ),
   },
-};
+} satisfies Story;
 
-export const NoPermissionRunning: Story = {
+export const NoPermissionRunning = {
   parameters: {
     ...withBuilds({ selectedBuild: inProgressBuild, userCanReview: false }),
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=2127-449276&mode=design&t=gIM40WT0324ynPQD-4"
     ),
   },
-};
+} satisfies Story;
 
-export const NoPermissionNoChanges: Story = {
+export const NoPermissionNoChanges = {
   parameters: {
     ...withBuilds({ selectedBuild: passedBuild, userCanReview: false }),
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=2127-449276&mode=design&t=gIM40WT0324ynPQD-4"
     ),
   },
-};
+} satisfies Story;
 
-export const ToggleSnapshot: Story = {
+export const ToggleSnapshot = {
   parameters: {
     ...withBuilds({ selectedBuild: pendingBuild }),
     ...withFigmaDesign(
@@ -480,9 +481,9 @@ export const ToggleSnapshot: Story = {
     const button = await findByTestId(canvasElement, "button-toggle-snapshot");
     await fireEvent.click(button);
   }),
-};
+} satisfies Story;
 
-export const Accepting: Story = {
+export const Accepting = {
   parameters: {
     msw: {
       handlers: [
@@ -500,9 +501,9 @@ export const Accepting: Story = {
     const button = await findByRole(canvasElement, "button", { name: "Accept" });
     await fireEvent.click(button);
   }),
-};
+} satisfies Story;
 
-export const AcceptingFailed: Story = {
+export const AcceptingFailed = {
   parameters: {
     msw: {
       handlers: [
@@ -520,18 +521,18 @@ export const AcceptingFailed: Story = {
       expect(argsByTarget["manager-api"].addNotification).toHaveBeenCalled()
     );
   }),
-};
+} satisfies Story;
 
-export const Accepted: Story = {
+export const Accepted = {
   parameters: {
     ...withBuilds({ selectedBuild: acceptedBuild }),
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=508-305053&t=0rxMQnkxsVpVj1qy-4"
     ),
   },
-};
+} satisfies Story;
 
-export const Skipped: Story = {
+export const Skipped = {
   args: {
     storyId: "button--tertiary",
   },
@@ -552,18 +553,18 @@ export const Skipped: Story = {
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=2255-42087&t=a8NRPgQk3kXMyxqZ-0"
     ),
   },
-};
+} satisfies Story;
 
-export const CaptureError: Story = {
+export const CaptureError = {
   parameters: {
     ...withBuilds({ selectedBuild: brokenBuild }),
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=508-305053&t=0rxMQnkxsVpVj1qy-4"
     ),
   },
-};
+} satisfies Story;
 
-export const InteractionFailure: Story = {
+export const InteractionFailure = {
   parameters: {
     ...withBuilds({
       selectedBuild: withTests(
@@ -575,26 +576,26 @@ export const InteractionFailure: Story = {
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=508-305053&t=0rxMQnkxsVpVj1qy-4"
     ),
   },
-};
+} satisfies Story;
 
-export const InfrastructureError: Story = {
+export const InfrastructureError = {
   parameters: {
     ...withBuilds({ selectedBuild: failedBuild }),
   },
-};
+} satisfies Story;
 
 /** The new build is newer than the story build (but we didn't run it) */
-export const CIBuildNewer: Story = {
+export const CIBuildNewer = {
   parameters: {
     ...withBuilds({
       selectedBuild: pendingBuild,
       lastBuildOnBranch: { ...pendingBuild, id: "2", committedAt: meta.args.gitInfo.committedAt },
     }),
   },
-};
+} satisfies Story;
 
 /** The new build is newer than the story build and the git info */
-export const CIBuildNewerThanCommit: Story = {
+export const CIBuildNewerThanCommit = {
   parameters: {
     ...withBuilds({
       selectedBuild: pendingBuild,
@@ -605,9 +606,9 @@ export const CIBuildNewerThanCommit: Story = {
       },
     }),
   },
-};
+} satisfies Story;
 
-// export const RenderSettings: Story = {
+// export const RenderSettings = {
 //   parameters: {
 //     ...withFigmaDesign(
 //       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=508-525764&t=18c1zI1SMe76dWYk-4"
@@ -617,9 +618,9 @@ export const CIBuildNewerThanCommit: Story = {
 //     const button = await findByRole(canvasElement, "button", { name: "Show render settings" });
 //     await fireEvent.click(button);
 //   }),
-// };
+// } satisfies Story;
 
-// export const Warnings: Story = {
+// export const Warnings = {
 //   parameters: {
 //     ...withFigmaDesign(
 //       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=516-672810&t=18c1zI1SMe76dWYk-4"
@@ -629,4 +630,4 @@ export const CIBuildNewerThanCommit: Story = {
 //     const button = await findByRole(canvasElement, "button", { name: "Show warnings" });
 //     await fireEvent.click(button);
 //   }),
-// };
+// } satisfies Story;