+++ src/screens/VisualTests/graphql.ts
@@ -128,13 +128,13 @@ export const FragmentStoryTestFields = graphql(/* GraphQL */ `
         version
       }
       captureDiff {
-        diffImage {
+        diffImage(signed: true) {
           imageUrl
           imageWidth
         }
       }
       headCapture {
-        captureImage {
+        captureImage(signed: true) {
           imageUrl
           imageWidth
         }
@@ -152,7 +152,7 @@ export const FragmentStoryTestFields = graphql(/* GraphQL */ `
         }
       }
       baseCapture {
-        captureImage {
+        captureImage(signed: true) {
           imageUrl
           imageWidth
         }